Port to Open Watcom 2.0: the game and tools build and run again
New build: build/watcom (wmake makefile, setenv.cmd, dosbox.cmd + dosbox.conf).
16-bit DOS, large model, 386 code, one translation unit, run in DOSBox-X.

src/compat.h maps the Borland-only calls (farmalloc, setvect, outportb,
random, delay, gotoxy, ...) to Open Watcom and provides M_PI.

Mechanical changes: asm{} blocks become _asm {}, <alloc.h>/<mem.h> replaced,
vga.c includes vga.h, Stock[] gets an explicit int type, the _AL pseudo
register is read through asm, asm labels made unique per file.

Borland behaviours that had to be reproduced:
- char is signed (-j): the code stores -1 in char (Sens, Rnd5/Rnd6, On_Obj)
  and without it every sprite and menu text was misplaced;
- Borland keeps DS on the data segment, Open Watcom lets DS float and reads
  globals through SS; every asm block that names a global (Video, TX,
  xms_driver, Cur_Pos) now loads DS from SS first. This was a random wild
  write of 63 KB (Drawy) that corrupted DOS memory within two menu frames;
- Borland's random() accepts a float argument: same macro reproduced;
- floats used directly as array indexes are cast explicitly.

Latent bugs fixed on the way (they only worked by luck of memory layout):
- start.c read Fich[3] past a 3-element array and fed stack garbage to the
  PCX loader;
- option.c read %d with fscanf into unsigned char variables and fields;
- cd.c wrote Tracks[-1] when no CD is present;
- plans.c Draw_Deg popped DS and ES in the wrong order;
- parameters named like globals (x, y, Page, Pal, pal) in asm functions
  renamed, since the inline assembler may bind the wrong symbol;
- credits.c had a stray */ inside an asm block.

tools/editpal.c never compiled (unfinished) and is not in the default target.
Verified in DOSBox-X: intro, main menu, Options and Teams screens.

+++ src/armes.c
@@ -36,7 +36,7 @@ if(Signe!=0)
 	temp=(float)Bombe.DirY/Bombe.DirX*Signe;
 	for(cont=0;cont<Bombe.DirX*Signe;cont++)   //Teste pixel par pixel...
 		{
-		if(Plan_1[(Bombe.PosX+cont*Signe)/320][Y[200-(Bombe.PosY+cont*temp)]+(int)(Bombe.PosX+cont*Signe)%320])
+		if(Plan_1[(int)((Bombe.PosX+cont*Signe)/320)][Y[(int)(200-(Bombe.PosY+cont*temp))]+(int)(Bombe.PosX+cont*Signe)%320])
 			{
 			Bombe.PosX+=cont*Signe;
 			Bombe.PosY+=temp*cont;
@@ -93,7 +93,7 @@ else
 	Bombe.PosY+=Bombe.DirY;
 	}
 if((Bombe.PosX>x)&&(Bombe.PosX<(x+320))&&(Bombe.PosY>200-y)&&(Bombe.PosY<400-y))  //si la bombe est dans l'ecran,...
-	Page[(Y[400-y-Bombe.PosY]+(Bombe.PosX-x))]=255;                                //la dessiner
+	Page[(unsigned)(Y[(int)(400-y-Bombe.PosY)]+(Bombe.PosX-x))]=255;                                //la dessiner
 }
 		 */
 
@@ -148,7 +148,7 @@ else                                                         //Si la bombe est a
 	Bombe.PosY+=Bombe.DirY;
 	}
 if((Bombe.PosX>x)&&(Bombe.PosX<(x+320))&&(Bombe.PosY>200-y)&&(Bombe.PosY<400-y))  //si la bombe est dans l'ecran,...
-	Page[(Y[400-y-Bombe.PosY]+(Bombe.PosX-x))]=255;                                //la dessiner
+	Page[(unsigned)(Y[(int)(400-y-Bombe.PosY)]+(Bombe.PosX-x))]=255;                                //la dessiner
 }
